fix(05_badusb): 修复 HID 注入问题, 支持中文输入法规避, 新增注入监控台
- hid_keyboard: tud_hid_report(1,..) report_id 参数错误导致松键被解析为 Ctrl 卡键; hid_key_tap 改用 hid_delay_ms 服务 USB 栈
- main: GPIO25 为 LED/LCD背光共用引脚, LED(1) 会关掉背光, 改用 LCD_PWR(0); hello.txt 不存在或与内置示例不一致时自动刷新
- payload: 支持纯修饰键组合, ALT SHIFT 切换键盘布局规避中文拼音 IME 组词乱码
- sample_script: 示例脚本内置 ALT SHIFT 前缀
- tools: 新增 rawkey_listener 目标机按键监听器 + monitor 网页监控台 (板子侧/目标机侧实时对照)
- 附带: 06/07 终端、公共库 pio_lcd/console/commands/tui、sdk_project.cmake、README 等未提交改动

+++ common/Middlewares/FatFs_SPI/sd_driver/sd_card.c
@@ -657,7 +657,7 @@ static uint32_t sd_go_idle_state(sd_card_t *pSD) {
         busy_wait_us(100 * 1000);
         sd_acquire(pSD);
     }
-    printf("response:%d,\r\n",response);
+    printf("response:%lu,\r\n",(unsigned long)response);
     return response;
 }
